+++ src/core/persistence/migrations.rs
@@ -1,1 +1,2 @@
 pub mod version_1;
+pub mod version_2;