+++ src/services/db.rs
@@ -29,7 +29,7 @@ pub fn spawn(service_id: &str, settings: &Settings, db: &Arc<Mutex<Connection>>,
         move || -> Result<()> {
             loop {
                 let size = select_size(&db.lock().unwrap())?;
-                tx.send(Composer::new(sensor.clone()).value(Value::Size(size)).into())?;
+                tx.send(Composer::new(sensor.clone()).value(Value::DataSize(size)).into())?;
                 thread::sleep(interval);
             }
         },